eureka

+++ notification-starter/src/main/java/ru/yandex/practicum/bank/notification/service/NotificationOutboxProcessor.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.bank.notification.service;
 
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestClient;
+import org.springframework.web.client.RestTemplate;
 import ru.yandex.practicum.bank.notification.dto.NotificationDto;
 import ru.yandex.practicum.bank.notification.mapper.NotificationMapper;
 import ru.yandex.practicum.bank.notification.model.Notification;
@@ -13,11 +13,11 @@ import java.util.List;
 @Service
 public class NotificationOutboxProcessor {
 
-    private final RestClient restClient = RestClient.create();
-
+    private final RestTemplate internalRestTemplate;
     private final NotificationJpaRepository notificationJpaRepository;
 
-    public NotificationOutboxProcessor(NotificationJpaRepository notificationJpaRepository) {
+    public NotificationOutboxProcessor(RestTemplate internalRestTemplate, NotificationJpaRepository notificationJpaRepository) {
+        this.internalRestTemplate = internalRestTemplate;
         this.notificationJpaRepository = notificationJpaRepository;
     }
 
@@ -26,15 +26,11 @@ public class NotificationOutboxProcessor {
         List<Notification> notifications = notificationJpaRepository.findAll();
         List<NotificationDto> notificationsDto = notifications.stream().map(NotificationMapper.INSTANCE::toNotificationDto).toList();
 
-        restClient
-                .post()
-                .uri("http://localhost:8088/notifications")
-                .body(notificationsDto)
-                .retrieve()
-                .toBodilessEntity();
-
-        notificationJpaRepository.deleteAll(notifications);
-
+        if (!notificationsDto.isEmpty()) {
+            internalRestTemplate
+                    .postForEntity("lb://notification-service/notifications", notificationsDto, Void.class);
+            notificationJpaRepository.deleteAll(notifications);
+        }
     }
 
 }
